+++ Stores/PostgreSQL/Cleipnir.ResilientFunctions.PostgreSQL/PostgreSqlMessageStore.cs
@@ -28,8 +28,8 @@ public class PostgreSqlMessageStore(string connectionString, SqlGenerator sqlGen
         _initializeSql ??= @$"
             CREATE TABLE IF NOT EXISTS {_tablePrefix}_messages (
                 id UUID,
-                position INT,
-                content BYTEA,         
+                position BIGINT,
+                content BYTEA,
                 PRIMARY KEY (id, position)
             );";
         
@@ -134,7 +134,7 @@ public class PostgreSqlMessageStore(string connectionString, SqlGenerator sqlGen
     }
 
     private string? _replaceMessageSql;
-    public async Task<bool> ReplaceMessage(StoredId storedId, int position, StoredMessage storedMessage)
+    public async Task<bool> ReplaceMessage(StoredId storedId, long position, StoredMessage storedMessage)
     {
         await using var conn = await CreateConnection();
         _replaceMessageSql ??= @$"    
@@ -179,7 +179,7 @@ public class PostgreSqlMessageStore(string connectionString, SqlGenerator sqlGen
         await command.ExecuteNonQueryAsync();
     }
 
-    public async Task<IReadOnlyList<StoredMessage>> GetMessages(StoredId storedId, int skip)
+    public async Task<IReadOnlyList<StoredMessage>> GetMessages(StoredId storedId, long skip)
     {
         await using var conn = await CreateConnection();
         await using var command = sqlGenerator.GetMessages(storedId, skip).ToNpgsqlCommand(conn);
@@ -221,11 +221,11 @@ public class PostgreSqlMessageStore(string connectionString, SqlGenerator sqlGen
         return storedMessage;
     }
 
-    public async Task<IDictionary<StoredId, int>> GetMaxPositions(IReadOnlyList<StoredId> storedIds)
+    public async Task<IDictionary<StoredId, long>> GetMaxPositions(IReadOnlyList<StoredId> storedIds)
     {
         if (storedIds.Count == 0)
-            return new Dictionary<StoredId, int>();
-        
+            return new Dictionary<StoredId, long>();
+
         var sql = @$"
             SELECT id, MAX(position)
             FROM {tablePrefix}_messages
@@ -235,18 +235,18 @@ public class PostgreSqlMessageStore(string connectionString, SqlGenerator sqlGen
         await using var conn = await CreateConnection();
         await using var command = new NpgsqlCommand(sql, conn);
 
-        var positions = new Dictionary<StoredId, int>(capacity: storedIds.Count);
+        var positions = new Dictionary<StoredId, long>(capacity: storedIds.Count);
         foreach (var storedId in storedIds)
             positions[storedId] = -1;
-        
+
         await using var reader = await command.ExecuteReaderAsync();
         while (await reader.ReadAsync())
         {
             var storedId =  new StoredId(reader.GetGuid(0));
-            var position = reader.GetInt32(1);
+            var position = reader.GetInt64(1);
             positions[storedId] = position;
         }
-        
+
         return positions;
     }
 }
